op-program: split into client and host packages
